Refactor (#21)
* Restructure models

* fix imports

* Update examples/nanoChatGPT/models/utils.py

* Additional restructuring

* Add example usage for reward model

* Rework RLHF class

* Additional cleanup

* refactor training loops

* further refactoring (#22)

* further refactoring

* some further changes

* Fixes

---------


---------
